StopVM to do forceful shutdown if VM is in paused state
StopVM calls an agent inside the VM. If the VM is paused StopVM can
be stuck as it cannot connect to the agent. To fix this, we implemented
api in go-sdk to read instance state and here we call it internally to
read the VM state. If it is paused then we do a forceful shutdown of the
VM.

Signed-off-by: Alakesh Haloi <alakeshh@amazon.com>